Rename chorus.ts and voice.ts to legacychorus.ts legacyvoice.ts respectively
